requirements.txt
ffmpeg-python==0.2.0
future==1.0.0
PyAudio==0.2.14
